Refactor: Participant 엔티티에 상태 코드, Enum 코드 추가
